Fixed bug in idClip in Translation of models not centered at origin.
This bug was introduced when idClip was migrated to octree:
  Revision: 9918
  Author: stgatilov
  Date: 3 May 2022 г. 7:36:48
  Message: #5954. Optimized spatial data structure in idClip.

Basically, we have an local origin point moving from "start", and a bounding box in the local system.
The center of the bounding box is as "queryStart": if this is same as "start", then everything is OK.
The problem is: we call octree.QueryInMovingBox with "queryStart" as initial point.
But later when we check individual clipmodels, we use "start" instead.


git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@10798 49c82d7f-2e2a-0410-a16f-ae8f201b507f