#6546. Added a way to make image CPU-resident during runtime.
The new function idImageManager::EnsureImageCpuResident check residency.
If image is not CPU-resident, then the image is reloaded with increased residency.

However, it is forbidden to call EnsureImageCpuResident directly in runtime.
We can't do any modifications to images or image manager during the frame, because backaend and frontend might use it in parallel!
That's why I added idImageManager::ExecuteWhenSingleThreaded function that defers execution of given lambda until the end of frame, where it is thread-safe.
This mechanism should be used for EnsureImageCpuResident, ImageFromXyz, and other potentially modifying functions.


git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@10817 49c82d7f-2e2a-0410-a16f-ae8f201b507f