#6546. Refactor loading particle collisions texture using new defer mechanism.
The problem here is when particles have static collisions, there corresponding textures are not referenced anywhere directly.
If we load the texture during runtime, then adding the image becomes race condition.
That's why I previously added hacky "preload" code in idRenderWorldLocal::ParseModel.

Now it is possible to load image from particle processing code, given that the load is delayed until the end of the frame.
Of course, we should be ready to get image = NULL and somehow work on this first frame.
The old hacky preload code is deleted.

The actual handling of defer stuff is in idParticleStage::LoadCutoffTimeMap now.
First we check if the image is already available with proper residency --- then we don't do anything special.
But if it's not available, then we load it --- either directly or via defer mechanism depending on caller's will.


git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@10818 49c82d7f-2e2a-0410-a16f-ae8f201b507f