#6546. Save/restore Light Estimate System.
The state of the system consists of two halves:
 * set of tracked entities + track duration are part of game state, we should save/restore them.
 * samples, models, etc. --- this all depends on models and thus is not part of game state.

So we only save the which entities were tracked and for which duration.
On restore, we read this information into special member "loadedEntities".
Then we consume it on next Think call: make these entities tracked and set appropriate tracked duration.

Note that we cannot apply information about loaded entities immediately because render entities are not set yet at restore time.


git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@10832 49c82d7f-2e2a-0410-a16f-ae8f201b507f