#6571. Now parallax effect properly modifies "to-light" and "to_view" vectors.
These vectors are used for shading equation and for shadow maps, including the self-shadows.
The last thing is most important actually.

Previously you could see how shadows move along the surface as you move camera around.
Now this effect should not happen, at least on perfectly flat surfaces.

As for non-planar surfaces... we might see some angles and non-smoothness in shadows now.
It could be avoided if the texcoord-vs-mesh scale was passed as vertex attribute.
But now I have to hack it out from screen-space derivatives, which is not very smooth sadly.


git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@10876 49c82d7f-2e2a-0410-a16f-ae8f201b507f