Added a way to force-disable tonemapping everywhere if 3D game is not running (r_tonemapOnlyGame3d).
This can be useful because the new tonemapping settings affect main menu rather negatively.
Well, in fact even the simple gamma already affects cinematics the bad way, introducing ugly color banding on dark colors.

Note: normally we should do tonemapping after 3D game is rendered but before GUI overlays are rendered.
But this is a hard topic, because some of the GUI overlays are not actually GUI (e.g. underwater overlay), so this requires some kind of manual tagging.
Better leave this for the future.

git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@10914 49c82d7f-2e2a-0410-a16f-ae8f201b507f